refactor after Anze recommendations
