+++ src/layer/S1GRDEOCloudLayer.ts
@@ -2,7 +2,8 @@ import { DATASET_EOCLOUD_S1GRD } from 'src/layer/dataset';
 
 import { AbstractSentinelHubV1OrV2Layer } from 'src/layer/AbstractSentinelHubV1OrV2Layer';
 import { AcquisitionMode, Polarization } from 'src/layer/S1GRDAWSEULayer';
-import { OrbitDirection, MosaickingOrder, Link, LinkType, RequestConfiguration } from 'src/layer/const';
+import { OrbitDirection, MosaickingOrder, Link, LinkType } from 'src/layer/const';
+import { RequestConfiguration } from 'src/utils/cancelRequests';
 
 /*
   Note: the usual combinations are IW + DV/SV + HIGH and EW + DH/SH + MEDIUM.
