+++ src/layer/S3OLCILayer.ts
@@ -1,11 +1,10 @@
 import moment from 'moment';
 
 import { BBox } from 'src/bbox';
-
-import { PaginatedTiles, Link, LinkType, RequestConfiguration } from 'src/layer/const';
-
+import { PaginatedTiles, Link, LinkType } from 'src/layer/const';
 import { DATASET_S3OLCI } from 'src/layer/dataset';
 import { AbstractSentinelHubV3Layer } from 'src/layer/AbstractSentinelHubV3Layer';
+import { RequestConfiguration } from 'src/utils/cancelRequests';
 
 export class S3OLCILayer extends AbstractSentinelHubV3Layer {
   public readonly dataset = DATASET_S3OLCI;
@@ -14,8 +13,8 @@ export class S3OLCILayer extends AbstractSentinelHubV3Layer {
     bbox: BBox,
     fromTime: Date,
     toTime: Date,
-    maxCount?: number,
-    offset?: number,
+    maxCount: number | null = null,
+    offset: number | null = null,
     reqConfig?: RequestConfiguration,
   ): Promise<PaginatedTiles> {
     const response = await this.fetchTiles(
