+++ src/layer/WmsLayer.ts
@@ -1,10 +1,11 @@
 import moment, { Moment } from 'moment';
 
 import { BBox } from 'src/bbox';
-import { GetMapParams, ApiType, RequestConfiguration } from 'src/layer/const';
+import { GetMapParams, ApiType } from 'src/layer/const';
 import { wmsGetMapUrl } from 'src/layer/wms';
 import { AbstractLayer } from 'src/layer/AbstractLayer';
 import { fetchGetCapabilitiesXml } from './utils';
+import { RequestConfiguration } from 'src/utils/cancelRequests';
 
 interface ConstructorParameters {
   baseUrl?: string;
